requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord==1.0.1
discord-sentry-reporting==2.0.4
discord.py==1.7.1
idna==3.1
multidict==5.1.0
numpy==1.20.2
opencv-python==4.5.1.48
PyYAML==5.4.1
sentry-sdk==1.0.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
